Skip casting if there is no change in type (#7147)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7147

In some expression with row type reordering, the planner generates a
cast expression that have the same type for input and output.  For large structs
it will still recurse into children and spend e.g. 27% of the CPU time doing
nothing.  Skip this step if we can detect the types are identical.

Reviewed By: mbasmanova

Differential Revision: D50457183

fbshipit-source-id: bf5a43ffce0a6963ef7873d8376eb77d0f13e8dc