Make the data directory used by QueryRunners in sync
HiveExternalWorkerQueryRunner and HiveQueryRunner originally can access
the data created by each other. However this useful property was lost
some time ago, due to a storage format string added in the middle of the
path. e.g. If the user pass "/Users/yingsu" as the DATA_DIR,
HiveQueryRunner creates the data at "/Users/yingsu/hive_data", but
HiveExternalWorkerQueryRunner creates the data at "/Users/yingsu/DWRF/
hive_data". This commit adds a "addStorageFormatToPath" flag to the
HiveExternalWorkerQueryRunner and set it to false. Other tests still
add it to the path for backward compatibility.