Prepare flat map struct encoding column reader for parallel decoding (#7155)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7155

We'll parallelize this loop with an executor in a future diff, so we're moving it to the end of the method to make it easier to reason that this method can return while the tasks initiated (in the future) in the loop are still running. The caller of the entire tree will have to wait though.

Reviewed By: Sullivan-Patrick

Differential Revision: D50472518

fbshipit-source-id: 0f3f3de1ebecc18aac6f1667eb25375d3784101e