Fix flaky raceBetweenRaclaimAndJoinFinish (#7227)
Summary:
The flakiness is due to the fact that if the last hash build operator is the first
child build memory pool in its parent or not. The hash build pool's parent
memory pool has a customized memory reclaim which only reclaim from the
first encountered hash build memory pool as hash build operator do coordinated
memory reclaim. The child pool visit is based on the child memory pool list which
is fixed on creation. The flaky test is designed to let the last build operator to
finish (the driver is closed and removed from the task), and block the remaining
hash build operators and the do memory reclamation in expect to reproduce the
corner case that all the hash build operators are in wait for build state and the
spiller are finalized. The test expect to see non-reclaimable counter to be set.

Given that if the last hash build is the first hash build pool in its parent pool's list,
then non-reclaimable counter  is zero, otherwise one.

This PR fixes this issue by checking whether the last build operator is the first hash
build pool and use the fact to check the test result. Run 100 times in clion and Meta
internal stress test.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7227

Reviewed By: tanjialiang, mbasmanova

Differential Revision: D50631838

Pulled By: xiaoxmeng

fbshipit-source-id: 6db51e933bff943ad1a6be577243166842d95781