Fix MapFromEntries corner case that generates invalid map by avoiding reusing the input array sizes. (#7268)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7268

Consider the input is an array vector with Array at index 0 being {(null), (null)} of size 2 and array at
index 1 being an empty array. What happens is the map vector generated for this will reuse the sizes of
the input which is [2, 0] and reuse the children of the input row vector which are the elements of the
array as keys and values. In this case the size of those is 0. Hence an invalid map is generated then
passed to checkDuplicateKeys which will throw VeloxRuntimeError during sorting.

The fix is to not reuse the sizes of the inputs and always allocate a new one with 0 sizes for
unselected rows.

fix https://github.com/facebookincubator/velox/issues/7241

Reviewed By: kevinwilfong

Differential Revision: D50676547

fbshipit-source-id: 4c71df649070c5e93b2afe6769a04f393cb026bf