[native] Convert input field names to lower case in plan conversion
Presto is case insensitive, and so always converts column names and field names
to lower case in its plan.

When Presto constructs the Velox query plan it passes in the types for Hive
tables as they were read from the metastore, with capital letters if they were
there in the metastore. Velox uses these types to construct the output type of
the TableScan.

This leads to errors as the output of the TableScan will have capital letters in
field names, while field accesses in the rest of the plan will use the field
names Presto used in its original plan, which are all lower case.

Converting the field names from the types in the input Hive tables to lower case
fixes this inconsistency.