[native] Add static mapping of native query configs
This change updates the mapping of native only query configs from a
prefix-based system to a hardcoded, static mapping. This is to
decouple naming conventions that can be different between presto
and velox. For instance, Velox permits the use of dots, whereas Presto
employs dots to indicate the catalog to which a property should
be linked.