Optimize remaining filter evaluation (#7260)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7260

Currently when we have a remaining filter separated by ANDs and same field is multi-referenced from different clauses, we eagerly materialize that field.  If there is a very selective condition in the AND clauses, we lose the opportunity to avoid reading unneeded rows using that condition on the mutli-referenced field.

Since AND always reduces the selectivity, there is no need to eagerly evaluate the multi-referenced fields.  It gives 2.6x performance gain in some queries by evaluating the field lazily.

Reviewed By: oerling

Differential Revision: D50700083

fbshipit-source-id: af24a4b1b9eb9857c805aabd7b77bdf85805a92e