Simplify the hash aggregation spill processing (#7317)
Summary:
Simplify the hash aggregation spill processing by removing the
fine-grained partition control logic in grouping set:
- Always create aggregation spiller with single partition
- Always spill all the rows from spiller no matter it is triggered by memory arbitration,
   threshold or test injection
- Simplify the corresponding un-spilling code path.

The final or single aggregation is multi-threaded executed on a worker which means
that each aggregate operator runs a subset of data so that fine-grained partitioning
inside an operator is not that necessary.

The followup is to remove the corresponding query configs and simplification inside
the spiller.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7317

Reviewed By: mbasmanova

Differential Revision: D50804640

Pulled By: xiaoxmeng

fbshipit-source-id: 45156936b6f070abb5e470b47ca4742396d0de73