Reuse buffers if executed by the same thread. (#7284)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7284

Reuse buffers by thread id.

In the single-threaded version we reuse the same buffer, which we even keep as a member variable, so we don't allocate on the next call.

In the multi-thread case we decided to save memory by not keeping a copy, but we can still share across threads so they're reused in lambdas executed by the same thread.

Reviewed By: Sullivan-Patrick

Differential Revision: D50748317

fbshipit-source-id: d1877e773708ebddf414b523b2eab6f79566a728