Remove invalid query ID cache from HBO cache manager
This cache adds a query ID when HBO failed to fetch history stats for
this query. Later during cost based planning, it will skip reading history
stats. However, the saving is small as the reading will be from cache.
And currently we run HBO optimizer twice, getting empty stats for one run
does not necessarily mean getting empty stats for the other run. This is
an over optimization and I remove it here.