fast path for storing a string into HashStringAllocator (#7355)
Summary:
Making a stream and appending a string and finishing a write to HashStringAllocator takes a long time.  The common use case is an append where there is one free list item that spans the rest of the last arena of the HSA. This makes a fast path for allocating the head of this fre block without first unhooking it from the free list and then putting the tail back there.

Using a test with all unique strings as in

select l_orderkey, l_comment, substr(l_comment, 6), count(*) from lineitem group by 1, 2, 3 order by 1 limit 1;

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7355

Reviewed By: Yuhta

Differential Revision: D50864970

Pulled By: oerling

fbshipit-source-id: a8246b1c93db639adb7137ffebae9add35901657