Fix tsan data race in flaky AllTableWriterTest.directReadWrite (#7424)
Summary:
In partitioned table write, table scan and table writer are in different pipelines.
The tsan test mode detects race condition on access to the output vector reused
by the table scan. The table scan reuse write to fields of the vector which were
previously read by the table write. This is expected and not a real data race and
table scan reuse code will check if the buffer is unique.
This PR fixes this by marking the fields detected by tsan as tsan_atomic so as to
de-flake the test case.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7424

Reviewed By: Yuhta, oerling

Differential Revision: D51016218

Pulled By: xiaoxmeng

fbshipit-source-id: a2c7565a35fa2c04470b9b1849becd147c7e9859