[native] Add caching of parsed Types
We've seen cases of queries that spend a large amount of time just parsing
types when converting the Presto Plan to Velox. This seems to be because it
parses the same large Row Types that are used across many field accesses.

Adding caching within a request shows a substantial decrease in the amount of
time it takes to do the conversion.

Notably, this helps with timeouts we're seeing making calls from the coordinator
to create tasks on the Workers.