Enforce queuing memory limit on exchange clients in merge exchange (#7410)
Summary:
Currently, when the merge exchange creates ExchangeClients for each
source, they are subject to a 32MB queuing limit. However, this can
lead to high memory usage if all clients queue around 32MB. For
instance, we observed 300 clients consuming over 6GB of memory in one
case. To address this issue, we have introduced a new query config
"merge_exchange.max_buffer_size" that sets an upper bound on the total
memory that can be queued by these clients. This max limit is divided
equally among all clients with an upper and lower limit of 32MB and
1MB, respectively, per client. The default for this config is set to
128MB. It's important to note that this limit is enforced
approximately, not strictly.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7410

Reviewed By: xiaoxmeng, mbasmanova

Differential Revision: D50996298

Pulled By: bikramSingh91

fbshipit-source-id: e41b994ea9c20f073d0a03115c5295738e9c0491