Optimize remaining filter evaluation (#7433)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7433

Currently when we have a remaining filter separated by ANDs and same field is multi-referenced from different clauses, we eagerly materialize that field.  If there is a very selective condition in the AND clauses, we lose the opportunity to avoid reading unneeded rows using that condition on the mutli-referenced field.

Since AND always reduces the selectivity, there is no need to eagerly evaluate the multi-referenced fields.  It gives 2.6x performance gain in some queries by evaluating the field lazily.

Reviewed By: laithsakka

Differential Revision: D51039406

fbshipit-source-id: 1cb0421b684332016014bbb9b3fc246c6055a593