Defer copying of non-key columns in TopN::addInput (#7172)
Summary:
Current implementation copies all the columns for the row that's added to the
priority queue to RowContainer. If a row is added to priority queue and then is
replaced by another row from the same batch of input, copying non-key columns
to RowContainer is wasteful.

This commit adds the following optimization. First, loop over the input rows,
identify rows that need to be added to the priority queue, copy key columns for
these to RowContainer and add these to the priority queue. For the input rows
that remain in the priority queue, copy non-key columns to the RowContainer.
This logic avoids copying non-key columns for rows that do not stay in the
priority queue.

Also, add a check for duplicate sorting keys to TopNNode's constructor.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7172

Reviewed By: Yuhta

Differential Revision: D51067442

Pulled By: mbasmanova

fbshipit-source-id: d64aade64ec550b8048c8ef049b865e4cef7c715