Refactor `buildTaskSpillDirectoryPath` to take `NodeConfig` directly.
We try to get `node.internal-address` or `node.ip` for building the task spill directory path even if `include-node-in-spill-path` is `false` (the default value is also `false`).

When `node.internal-address` and `node.ip` are not set, all queries will fail. This refactor will prevent the check of `node.internal-address` and `node.ip` if `include-node-in-spill-path` is `false`.