[native] Resolve antlr dependency for CTest on Apple Silicon
On Apple M1 antlr is installed to /usr/local/lib. When running Ctest (e.g. make unittest)
some tests fail due to antlr library not being found.
Executing the tests manually on a terminal works as expected. The issue is only in Ctest.
Setting the DYLD_LIBRARY_PATH explicitly in the environment does not resolve the issue and
the CMakeLists.txt for each affected test needs to be modified.