Add a way to override current time in tests (#7543)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7543

Adding a ScopedTestTime class that allows us to override the value returned by
getCurrentTimeMs/getCurrentTimeMicro in tests.

This will be useful for testing some of the times we set in Task for example by making them deterministic.

It only applies in debug mode so there's no risk of breaking or impacting performance in prod.

Reviewed By: xiaoxmeng

Differential Revision: D51267554

fbshipit-source-id: da4a81b9de566da86744b09e7b9e33478b49a64d