[Iceberg] Add changelog table
This commit exposes a new system table called `$changelog` that
allows users to return a table containing information about when
which snapshots specific records were added or deleted (or updated).

The schema of the changelog table is static (except for the `rowdata` column):

```
| operation: VARCHAR | ordinal: bigint | snapshotid: bigint | rowdata: row(<parent table schema>) |
```

`operation` is the type of operation. This will be `DELETE`, `INSERT`,
or `UPDATE`. The ordinal is a value which represents the order in
which the change should be applied relative to all the other changes
in this particular scan of the changelog. These values always start at
0 and increment for each change to the table. The snapshot ID refers
to the snapshot of the table for which the particular operation belongs
to. The rowdata is the row pertaining to the particular operation.

I've seen for CoW tables that updating a row with spark in the changelog
actually results in a `DELETE` and `INSERT` being encoded in the same
ordinal.

To query the changelog table, users simply need to provide the starting
snapshot ID for which they would like to query. The table name should be
in the format:

```sql
... FROM "<table_name>[@<begin snapshot>]$changelog@[<end snapshot>]"
```

The below query is an example of how the changelog table can be used
to view the time that a particular record was inserted

```

SELECT committed_at, change.operation, rowdata.orderkey as orderkey, ordinal
FROM "orders$snapshots" as snap
JOIN "orders@5922410230513748202$changelog" as change
ON change.snapshotid = snap.snapshot_id LIMIT 3;

                committed_at                 | operation | orderkey | ordinal
---------------------------------------------+-----------+----------+---------
    2023-09-21 16:36:49.098 America/Los_Angeles | INSERT    |    18019 |       0
    2023-09-21 16:36:49.098 America/Los_Angeles | INSERT    |    18018 |       0
    2023-09-21 16:36:49.098 America/Los_Angeles | INSERT    |    18017 |       0
```

At a high-level the implementation strategy needed to expose information
from Iceberg's `IncrementalChangelogScan` tasks which differs from a
normal `TableScan`. The plan tasks for incremental scans include the
information on operation type, ordinal, and snapshot ID which are are not
contained in the normal table scan plan.

When the connector metadata queries for a changelog table type, it
returns a schema with a static 4 columns, where the first 3 are static
VARCHAR, BIGINT, BIGINT, and the final is a row type whose inner types
are dependent upon the queried table.

To accomplish this the following changes were made:
- logic to generate the changelog table schema
- a new piece of info on iceberg splits containing the original table's
column handles which is used during split sourcing
- a new page source which constructs the correct format page based on
the original table schema
- finally, some small refactoring of the IcebergNativeMetadata and
IcebergHiveMetadata classes to consolidate certain bits of logic for the
ConnectorMetadata

In addition to the changelog implementation, a new function was added which
coincides with the changelog table called `apply_changelog`. This function
uses the changelog table columns as inputs in order to apply the changelog
to coalesce a potentially large number of changes over time into a single
row per group.

For example, imagine a row with primary key 12345. It may be inserted initially
at ordinal 2, updated in ordinal 3 and 4, and then deleted in ordinal 5. The
`apply_changelog` function would coalesce these into a final result showing
the final value for the row data - `NULL` in this example since the final operation
was a delete. Generally, this means the query using the `apply_changelog`
function will also have a `GROUP BY` clause where the grouping key will be a
primary key or column with unique values.

An example usage may be helpful:

```SQL
SELECT rowdata.orderkey, apply_changelog(ordinal, operation, rowdata order by ordinal) as rowdata
FROM "orders@5922410230513748202$changelog"
GROUP BY rowdata.orderkey
```

Co-authored-by: xiz675 <32505316+xiz675@users.noreply.github.com>