Fix generic in-predicate for complex types (#7545)
Summary:
This fixes a bug where the generic in-predicate, which is employed
when the in-list is not a constant, can return null if the input
contains a null even though the in-list is empty. The correct
behavior is that it should throw an error for the empty in-list.

Fixes https://github.com/facebookincubator/velox/issues/7533

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7545

Test Plan: Added unit test

Reviewed By: mbasmanova

Differential Revision: D51277276

Pulled By: bikramSingh91

fbshipit-source-id: 65373617327da0a3432dbc22c5c7bf4fd34ab2b8