Fix signatures of the IN Presto predicate (#7585)
Summary:
A natural signature of the IN predicate is (T, T,...) -> boolean. It takes 2 or
more arguments of the same type and returns a boolean, e.g. c IN (x, y, z).

When in-list is constant and long (dozens of values) this signature is
inefficient to execute. Each item in the in-list becomes its own
ConstantTypeExpr and requires evaluation and production of shared pointers to
constant vectors. For this reason, the signature for constant in lists is
(T, array(T)) -> boolean.

Using same signature for non-constant in-list is inefficient as it requires the
appliction to wrap in-list columns in an array_constructor, which is slow
because it needs to copy all the data.

  a IN (x, y, z) -> in(a, array_constructor(x, y, z))

Hence, we use (T, array(T)) -> boolean signature for constant in-list and
(T, T,...) -> boolean for non-constant in-list.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7585

Reviewed By: xiaoxmeng

Differential Revision: D51342602

Pulled By: mbasmanova

fbshipit-source-id: df70ba2d17abeeb287a2e0a56f7a2b3f6e0cabb3