Optimize PartitionedOutput to eagerly flush data if it is after partial limit (#7635)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7635

When `PartitionedOutput` is after a partial limit, it mean the consumer
side of the shuffle is a final limit, and will stop pulling data before we
process all the upstream.  In this case we should flush to the output buffer as
soon as possible so that the final limit can stop pulling data as early as
possible.  In some extreme cases we reduce the query CPU usage by more than 200
times (18.47 hours to 4.73 min).

Reviewed By: mbasmanova

Differential Revision: D51429641

fbshipit-source-id: f53192bea02447e88e8e8dc551ee4dade41b14cd