Allow lazy creation of spill directory (#7660)
Summary:
Currently the caller creating a task needs to ensure that the spill
directory is created before setting it in the task. This means that
a spill directory is created for each task even if it never required
(if spilling is not triggered). This can be wasteful as the create
and delete calls to the filesystem have a cost. This change adds the
ability to lazily create the directory on first use in a thread safe
way.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7660

Test Plan: Add a unit test

Reviewed By: xiaoxmeng

Differential Revision: D51485886

Pulled By: bikramSingh91

fbshipit-source-id: bf846d9baf71ea96b30473b8b075936743d69577