Fix fuzzer option containerHasNulls meaning (#7667)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7667

containerHasNulls used to control whether top-level container values can be
null, and the container elements themselves, as the name suggests. With this
change, I removed the reference to the flag to callsites where nullRatio is
already zero (and hence this is meaningless), check that other callsites are
covered and work well with the new meaning, and implemented the new semantic
described above.
Added unit tests to ensure it works as expected.

Reviewed By: xiaoxmeng, mbasmanova

Differential Revision: D51487492

fbshipit-source-id: 2e800da9f4981c24e58997ff882143b7d72f7b7c