Do not use system curl to build cpr (#7718)
Summary:
In https://github.com/facebookincubator/velox/issues/7385 we introduced `libcpr` to Velox. It set `CPR_USE_SYSTEM_CURL`
true, and it may cause a construction issue if the host installed a lower version curl.
This PR disables this flag to force the use of bundled curl in `libcpr`, to avoid
the potential curl version issues, and make the dependencies version more
accurate and controllable and isolated from the state of the host machine.

Also, see the blog https://velox-lib.io/blog/velox-build-experience.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7718

Reviewed By: mbasmanova

Differential Revision: D51630526

Pulled By: kgpai

fbshipit-source-id: 457b2270f222043291a100afe9f5991c8035dcb5