Add driver thread debug context when enqueue a closed driver (#7786)
Summary:
We found some time the driver thread debug context is not set on server
crash if it is triggered by driver destruction throws when we enqueue a
closed driver.
This PR adds the driver thread debug context when we enqueue a driver
to make sure any exception case is covered.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7786

Reviewed By: mbasmanova

Differential Revision: D51650495

Pulled By: xiaoxmeng

fbshipit-source-id: ab8d187fde691f86888ea570c6c5a546257f0e82