Optimize simd::fromBitMask (#7827)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7827

In `fromBitMask` we currently use a static LUT which requires LOCK prefix in each invocation, slow down the code and possible preventing it from inlining.  Optimize the LOCK prefix away so that this function shall no longer be visible in profile.  The key takeaway is that we should prefer global instead of `static` in case the function itself is simple.

Micro-benchmark shows improvement about 40x in throughput.  In some real world queries this sometimes shows up as a small percentage of CPU (~1%).

Reviewed By: mbasmanova

Differential Revision: D51735444

fbshipit-source-id: 50b2ea0c54b1a60a92e461d5bd8a31e7c3021d12