Fix flaky shared arbitrator tests (#7836)
Summary:
Concurrent arbitrator test discovered some race condition in hash build operators
that can lead to wrong task failure. For instance, when the last hash build operator
failed to build the table, it will leave the hash join bridge in a inconsistent state that
cause the other build operators fail when reach to the join bridge. Since these are
running in parallel, it can lead to incorrect the task failure if the other hash build
operators set the task error first. With a temporal fix, the test can easily get into stuck.
This PR disable the concurrent test and re-enable after all the fixes are in.
This PR also simplify an existing test case.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7836

Reviewed By: bikramSingh91

Differential Revision: D51759996

Pulled By: xiaoxmeng

fbshipit-source-id: 5b6eb4af9d283da3be1bdc6b6468f679314a0ef7