Enhancing readme for first time users
A common problem appears to be that first time users
copy the VM options into the Program arguments field.
This PR addresses the issue that the VM options field
may not visible by default.

In addition, using Java 11 or later in IntelliJ can
cause build issues. A certain option that is enabled
for the Java compiler by default needs to be disabled.