[native-pos] Throw fatal exception native process cannot find port
Currently, when we cannot find port to start the native
process, we mark only the task failed and not indicate that
anything is wrong with the executor itself. This results in
spark driver scheduling more tasks that fail with same error
ultimately leading to query failure due to one bad executor.

This commit throws a PrestoSparkFatalException indicating that
this executor crashed and create a new one. A new executor might
still land on a bad host and have the same issue, but we would
like to rely on spark's scheduler to detect a bad host. That
is beyond the scope of this commit