[Native] Don't propagate errors in getResults
Today, the native TaskManager will throw an exception if
getResults is called on a failed or aborted task. In Presto Java,
it looks like the Worker will return empty results rather than an
error in these cases.

Looking at ClientBuffer in the Java code, it creates a future for
getResults that will be fulfilled when the Task enqueues results.
If the Task does not enqueue results, e.g. because it's failed or
aborted, the future is fulfilled with empty results, either when
getResults is called again or when the ClientBuffer is destroyed.

This behavior makes sense to me.  By propagating the exception,
there's a race condition in the Cooridinator resulting in the
actual exception that caused the Task to fail, depending on
whether the producer or consumer Task is marked as failed first
in the Coordinator. If the Task is failed/aborted then the
Coordinator should eventually realize this through the
getTaskStatus call. Put another way, the purpose of getResults is
to propagate data, and the purpose of getTaskStatus is to
propagate errors.