Fix bug in tracking partial aggregation stats when the query is a group-by on the partitioning key
During tracking of histories, we adjust the output byte count to account for hash variables introduced during planning.
In certain cases, the reported byte count is 0 (for example when the grouping key is a partition key, its value does not contribute to the byte count). After accounting for hash variables, the 0 becomes a negative number,
which raises a NaN exception that looks like this:
java.lang.IllegalArgumentException: value is NaN
	at com.facebook.presto.spi.statistics.Estimate.of(Estimate.java:54)
	at com.facebook.presto.cost.HistoryBasedPlanStatisticsTracker.constructAggregationNodeStatistics(HistoryBasedPlanStatisticsTracker.java:250)
	at com.facebook.presto.cost.HistoryBasedPlanStatisticsTracker.getQueryStats(HistoryBasedPlanStatisticsTracker.java:164)

This PR fixes that by turning NaN values into unknown like it is done in a few other places of the HBO tracker code.