Add option to add limit to number of groups for khyperloglog_agg function
khyperloglog_agg create many hyperloglog objects during aggregation, and this
can lead to native memory usage for GC algrithm due to cross reference between
heap memory regions. This PR adds an option to add limit on the number of groups
for this aggregation function. Defaults to 0 which means no limit.