Fix outputBytes/Rows/Batches stats for PartitionedOutput (#8072)
Summary:
PartitionedOutput operator used to report outputBytes/Rows/Batches the same as
inputBytes/Rows/Batches. With this change it reports outputBytes as number of
serialized bytes, outputRows the same as inputRows and outputBatches as number
of serialized pages produced. This helps get an accurate picture while debugging
issues with Exchange.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8072

Reviewed By: xiaoxmeng

Differential Revision: D52207253

Pulled By: mbasmanova

fbshipit-source-id: 5f6b93cda6e5c7043b4ac77756187a45b9063fdc