Add maxSpillRunRows config (#7928)
Summary:
We have seen memory allocation failure from the memory allocator
under Prestissimo on Spark workload. The allocation size is very large ~1GB
and the allocation is made from a spilling memory pool. It is likely caused
by extremely large number of rows so the spill run vector needs to allocate
a large amount of memory. To limit the spilling memory usage, we would
better put a limit on the max spill run (a disk spilling config), that is, the
maximum number of rows in a sorted run on disk.

Details in https://github.com/facebookincubator/velox/issues/7760

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7928

Reviewed By: tanjialiang

Differential Revision: D52192397

Pulled By: xiaoxmeng

fbshipit-source-id: 7538cc012ddc38401e9121cf7987f519e99bb017