Allow connector derived from Hive to specialize filter pushdown
Meta's internal FilterPushdown inherits the HiveFilterPushdown. Due to the
refactoring on HiveFilterPushdown, Meta's Internal build is broken. In order
to fix the build, the methods in HiveFilterPushdown needs to be moved around
and changed for visibility in order for the derived class to use the methods
in the super class.