Reorder IndexJoin and AddNotNullFiltersToJoinNode optimizations to avoid index join optimization
Enabling the property joins_not_null_inference_strategy caused a regression for queries that were supposed to get evaluated using an index join.
The problem was that AddNotNullFiltersToJoinNode added additional predicates to outer joins, which caused IndexJoinOptimizer to not trigger, as
additional predicates in outer joins are not supported.

This fix reorders the optimization rules to avoid this issue.