Track the file ages for AyncDataCache and SsdCache (#6381)
Summary:
Use the file open time to calculate the age of AsyncDataCache
and SsdCache entries that are loaded from the file. Tracking this
enables us to expose the metrics maxAge possible of cache entries.

Add a process-wide singleton CacheTTLController to handle the TTL
of AsyncDataCache and SsdCache and track the required info. It
currently adds the file open time when a new file handle is opened. It
also calculates the maxAge for cache entries.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6381

Reviewed By: spershin

Differential Revision: D48959713

Pulled By: gggrace14

fbshipit-source-id: 4924d507cf851692b3b2e4d11f0c05fabcbf8475