Allow connector optimizers to create joins
The current join nodes are in presto-main. Moving them requires moving a
lot of objets.

Since we don't need to allow the connectos to optimize the joins for now
we can just add an adapter node that will be converted back to the
internal node type when returned by connector optimzers.