Make ScopedTestTime compile outside of debug mode (#8130)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8130

Wrapping the entire ScopedTestTime class definition in a #ifndef NDEBUG macro makes it so that tests using it don't build
outside of debug mode.

The intent was that ScopedTestTime shouldn't be used outside of debug mode.  Since the implementations of
getCurrentTime that use the class are already wrapped with the macro to protect production, we can achieve the same intent
by throwing an exception if the class is constructed outside of debug mode. This way, the test still compiles outside of debug
mode.

Reviewed By: xiaoxmeng

Differential Revision: D52340499

fbshipit-source-id: 3b38b6300a869d8e6e89185401d0b9ae88e75a33