Fix presto null handling in comparison operations. (#7703)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7703

Velox used to depend to use the StopAtNull mode to implement presto semantics
of comparisons. However as discussed in https://github.com/facebookincubator/velox/issues/7536
that mode does not correctly implement presto semantics.

This diff replaces StopAtNull mode with NullAsIndeterminate mode. NullAsIndeterminate
implements presto semantics for comparisons, the diff contains a detailed documentation of the
mode. But here I summarize how it differs from StopAtNull.

1) When equalsOnly=true, for complex types,  StopAtNull stops at first null encountered,
while NullAsIndeterminate checks all children and if any is false, the final result is false.
ex: [null, 1] = [null, 2] is false under NullAsIndeterminate.

2) When equalsOnly=false, if any null is encountered, NullAsIndeterminate throws a user error
"Ordering nulls is not supported" while StopAtNull returns std::nullopt. And callers used to
 manually handle that.

3) When equalsOnly=false comparing maps will throw a runtime error, "Map is not orderable
 type" while StopAtNull used to allow it.

This changes the behavior of several existing functions some of which are tested in this diff such as
array_remove, contains and ==.

Reviewed By: mbasmanova

Differential Revision: D51518489

fbshipit-source-id: 3931afb66bb3c707822485fa5ae029db07588ec1