Reduce link parallelism to fix compilation OOM's (#8403)
Summary:
**What ?**

We reduce the high memory jobs parallelism from 8 to 4 and the max number
of debug threads to 4 for the linux-pr-fuzzer run.

**Why ?**

This is to fix https://github.com/facebookincubator/velox/issues/8401 where we are seeing intermittent compilation failures.

**Notes:**

This change only affects linux-pr-fuzzer-run and not all the other jobs as they
have been succeeding with the level of parallelism. Why linux-pr-fuzzer-run
only fails could be due to additional compile time flags
(-DVELOX_ENABLE_ARROW=ON) that might be adding to the load. We will do a full
fix as a migration of these jobs to Github Actions from CCI.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8403

Reviewed By: mbasmanova

Differential Revision: D52823273

Pulled By: kgpai

fbshipit-source-id: 713e3f53fcbfd77926036cbaf469c58b6f8b8138