Optimize computation of partition start rows in SortWindowBuild (#7580)
Summary:
Before this change, computePartitionStartRows used linear search with time
complexity O(n) to find the partition start rows. In many cases where the
partitions have lots of rows this algorithm is inefficient. It can be optimized
using binary search, but the time complexity of back-to-back binary search is
worse when there are many partitions with few rows each. Forward-to-back double
search algorithm is more efficient in all cases.

The benchmark shows average improvement of 3% and the best case improvement of
20%.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/7580

Reviewed By: amitkdutta

Differential Revision: D52823941

Pulled By: mbasmanova

fbshipit-source-id: 16011a87e7d5897d4128ece4f1bb0c5f7f2afab5