Use same max-output-buffer-size config for all types of output buffers (#8426)
Summary:
We need 2 configuration settings to control the behavior of PartitionedOutput
operator.

First, a limit on how many bytes to buffer inside the PartitionedOutput operator
itself. That limit is split N ways (where N is number of partitions) to compute
per-destination packet size. A packet here is a SerializedPage.

Second, a limit on how many bytes to buffer in OutputBuffer across all
destinations (a sum of all ready packets / SerializedPages). This limit needs
to be significantly higher than the first limit.

We have introduced the second limit earlier to use for arbitrary buffer, but it
should be used for all kinds of buffers (partitioned, broadcast and
arbitrary).

This change is to start using this second limit for all kinds of buffers. A
series of follow-up changes needs to rename the configuration property to
remove 'arbitrary' from the name in a backwards compatible way
(max_arbitrary_buffer_size -> max_buffer_size).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8426

Reviewed By: spershin

Differential Revision: D52857143

Pulled By: mbasmanova

fbshipit-source-id: cf790e48d02652b45f2c46bc683ba0301dba323d