Fix bucket counts in histogram metrics (#8458)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8458

I noticed Velox's histogram metrics were consuming ~1 GB of memory.

Looking at the comments on the metrics, the authors tried to create the metrics with a small number of buckets, but the argument where they put the number of buckets is actually the width of buckets, so these metrics frequently have tens of thousands of buckets.

This fixes those metrics setting the bucket width so that the number of buckets created matches the comments.

Reviewed By: Yuhta

Differential Revision: D52927472

fbshipit-source-id: 589f10680120773c5dd672a9338d3bc13248fde0