Fix ExchangeBenchmark and PrefixSortAlgorithmBenchmark (#8493)
Summary:
terminate called after throwing an instance of 'facebook::velox::VeloxRuntimeError'
  what():  Exception: VeloxRuntimeError
Error Source: RUNTIME
Error Code: INVALID_STATE
Reason: The memory manager is not set

The reason for the error is that the benchmark object is initialized before the main function is called. We need to ensure that the memory manager instance is initialized before the benchmark object.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8493

Reviewed By: Yuhta

Differential Revision: D53042869

Pulled By: mbasmanova

fbshipit-source-id: f0abec44afa347702a444854396fbdf2a96bf195