Signal timeout when locking task in getLongRunningOpCalls (#8511)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8511

If there are some problem (e.g. deadlock) with the mutex in `Task`,
`getLongRunningOpCalls` would block forever waiting for it.  In this case we
should signal the error and let the caller to generate some form of alerts.
This is important as we rely on `getLongRunningOpCalls` to detect blocking and
the function itself should not block.

Reviewed By: xiaoxmeng

Differential Revision: D53048088

fbshipit-source-id: 5825ee5887a33ffcc97ab0acd54a3c6bab059441