Add memory reclaim stats in TaskStats (#8497)
Summary:
Add memory reclaim stats (time and count) in TaskStats to help analyze how much time
a task spent on memory reclaim by memory arbitration. If the memory arbitration is initiated
by the query itself. This time will be counted in its operator's  execution time. If not, the task
will be stopped for that long to reclaim memory. This indicates if a slow query is caused by
the background memory arbitration.

The followup is to report the stats as task runtime stats in Prestissimo.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8497

Reviewed By: mbasmanova

Differential Revision: D53035172

Pulled By: xiaoxmeng

fbshipit-source-id: 8f80981d0bd153a19ba3e517d60e41780bac279a