[native] Add memory reclaim stats to task runtime stats
Add task runtime stats to collect memory reclaim stats during a task execution.
This helps to debug how much time spent on task memory reclaim of a slow query
execution. If the memory reclaim is triggered by the task itself, then the time will
also be counted in task cpu execution time. If not, it is counted in the task's scheduled
time but not execution time as the task will be stopped before memory reclamation.