Eliminate redundant Expr::addNulls calls (#8616)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8616

Expr::addNulls can be expensive as it my require making vector writable by
copying it into a new vector. Hence, it is important to make sure not to call
it if there are no nulls to add.

This change avoids calling Expr::addNulls unnecessarily by tightening checks for
new nulls in MutableRemainingRows::mayHaveChanged and Expr::removeSureNulls.
Rename MutableRemainingRows::mayHaveChanged to hasChanged to match new
behavior.

This optimization reduces expression evaluation CPU and wall time 2x for one of
the affected queries.

This change uncovered a bug in tranform_keys Presto lambda function. The check
for keys containing nulls was incorrectly skipped if there were no null keys. This change
includes a fix.

Reviewed By: bikramSingh91

Differential Revision: D53264651

fbshipit-source-id: 18f4e9737f544c89cf35d3261e60dc93484cd8af