Allow offline pre-compilation of 'xsimd' dependency (#8308)
Summary:
`set(xsimd_SOURCE BUNDLED)` will force the 'xsimd' library to be
compiled from source, but this will make it impossible to compile Velox
offline. We should use `set_source(xsimd)` so that it respects the
`VELOX_DEPENDENCY_SOURCE` global cmake variable.

If it is really needed to force the 'xsimd' library to be compiled from
source in a certain environment, use the following method to achieve it.

```
cmake -Dxsimd_SOURCE=BUNDLED ...
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8308

Reviewed By: Yuhta

Differential Revision: D53320172

Pulled By: kgpai

fbshipit-source-id: 8f9fb975ef33ab9596facd1283151f060de545b7