Add support for deserialization of all-null columns serialized as UNKNOWN type (#8355)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8355

This change enables the deserialization of a column consisting
entirely of null values, regardless of its data type, when it is
serialized as a column of the UNKNOWN type. Such a vector isn't
produced by the PrestoSerializer, but it can be encountered when
processing vectors not generated by it. For instance, this situation
arose when a constant vector, serialized within the query plan and
created by the Presto coordinator, was being read.

Reviewed By: kevinwilfong

Differential Revision: D52711814

fbshipit-source-id: 29c6478bae055cc9e0663e13940be639726d01fd