Add new APIs to AddressableNonNullValueList to copy/append a stream of bytes (#8653)
Summary:
This is the first in a set of PRs to add support for spilling distinct aggregations (see full version in https://github.com/facebookincubator/velox/pull/7791).

Spilling distinct aggregations needs support to spill SetAccumulators in which input values are cumulated. ComplexTypeSetAccumulators use AddressableNonNullValueList to serialize complex types. This PR adds new APIs to AddressableNonNullValueList so that it can copy/append a stream of bytes corresponding to a ComplexType value (array, map, struct).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8653

Reviewed By: Yuhta

Differential Revision: D53497000

Pulled By: mbasmanova

fbshipit-source-id: 66d44d02a2c3bd5775725c8b8559feaed17c0813