Fix arbitrary() to always return the first non-null value in Window operation (#8640)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8640

The implementation of arbitrary() intends to always return the first non-null value, but it
doesn't for complex-typed inputs when used in Window operation with incremental frames.
This is because ArbitraryFunction::addSingleGroupRawInput() still updates the accumulator
even if the accumulator already has a value. This diff fixes this issue by making
ArbitraryFunction::addSingleGroupRawInput() return immediately if accumulator already has
a value.

This diff fixes https://github.com/facebookincubator/velox/issues/8593.

Reviewed By: kgpai

Differential Revision: D53328253

fbshipit-source-id: 803261dce0ec1fc52187947b1f9316dfd814c3fd