Move header & format check to Github Actions (#8546)
Summary:
This is the first in a number of PRs to move all of our CI over to GHA.

The format and header checks will now produce a markdown summary that is rendered in the job summary. This includes a code block with copy2clipboard-button for easy application of the required changes without looking through the logs.

Closes https://github.com/facebookincubator/velox/issues/8548

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8546

Reviewed By: kagamiori

Differential Revision: D53862082

Pulled By: kgpai

fbshipit-source-id: c85dc39c268b3369c0f5da1597a23a9d3d110b9c