Fix variable width encoding description in serialized-page.rst
The bullet points under "VARIABLE_WIDTH Encoding" list "total number of bytes in all values" as 4 bytes, but in the descriptive paragraph below it says 1 byte.  I checked the code and the correct value is 4 bytes.

I updated the text, but the image is wrong as well, I don't know how to regenerate it.