Remove unused variables in velox/expression/CastExpr-inl.h (#8773)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8773

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D53779591

fbshipit-source-id: 67b39c9f26a6cd6f783e7d621551cbfe158fe434