Remove unused variables in velox/common/memory/MmapArena.cpp (#8786)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8786

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D53779573

fbshipit-source-id: 82b9566721808b5f8596d067864a61dba6a52f3e