Remove unused variables in velox/functions/lib/aggregates/DecimalAggregate.h (#8789)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8789

LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D53779487

fbshipit-source-id: 78f3ba4b64051b09a5e1e5e45af867ecc948c62f