Enable `-Wimplicit-fallthrough` in f5b/PACKAGE (#8799)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8799

This diff enables the titular warning flag for the directory in question. Further details are in [this workplace post](https://fb.workplace.com/permalink.php?story_fbid=pfbid02XaWNiCVk69r1ghfvDVpujB8Hr9Y61uDvNakxiZFa2jwiPHscVdEQwCBHrmWZSyMRl&id=100051201402394).

This is a low-risk diff. There are **no run-time effects** and the diff has already been observed to compile locally. **If the code compiles, it works; test errors are spurious.**

If the diff does not pass, it will be closed automatically.

Reviewed By: dmm-fb

Differential Revision: D53545504

fbshipit-source-id: f1f54e6d085e79358b07696dbdb8e4b7fdb291cf