[native] Use PrestoSerializer to deserialize base64 encoded constants in presto plan
At present, Presto native has its own method for deserializing a
constant vector in the Presto plan it receives from the coordinator.
This plan is serialized in the Presto page format, excluding the
Presto page header. A recent bug was identified in this method,
revealing that it doesn't handle nulls in a top-level struct column.
On the other hand, Velox already provides a more robust and
well-tested implementation of the PrestoPage format, which is
employed for both serialization and deserialization. Consequently,
instead of rectifying the aforementioned bug, this PR eliminates
the redundant Presto page deserializer in the Presto native codebase.
It then integrates this with a call to the PrestoSerializer in Velox.

Test Plan:
Added additional unit test and passed existing.