Quick stats
Introduces a quick stats provider that builds quick stats from a
provider implementation. Stats can be built either in-line or in
the background. Computed stats are cached in-memory for future use

The time to build stats is kept bounded using a config property
A ParquetQuickStatsBuilder is also added, which computes stats
using parquet file footers