Encode the minimum necessary values from a dictionary in PrestoSerializer (#8639)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8639

There are cases where the values in a DictionaryVector may be much larger than what's actually needed.  E.g. some UDFs
use DictionaryVectors to filter values, like array subscript.  When we use PrestoBatchSerializer in PartitionedOutput with a
single destination, I suspect we'll run into similar issues if we have a large RowVector that we break into smaller pieces, we
don't necessarily want to send the whole alphabet for each piece.

This change adds logic to serializeDictionaryVector to determine the minimum subset of the alphabet to include in the
serialized output.  It also updates the indices to point to the correct values in this minimal dictionary.

Reviewed By: bikramSingh91

Differential Revision: D53327612

fbshipit-source-id: 1651554806d2124d982cca122fd707fcc84cab95