Add SelectivityVector to DecodedVector::nulls() (#8579)
Summary:
If DecodedVector decodes a partial set of rows and makes a local materialization of indices, nulls() without SelectivityVector will access all elements in indices, which will hit random memory locations. Therefore, pass a SelectivityVector to nulls(). Require that if decode() is for partial rows, then nulls() is also for partial rows. There The rows for null() must be a subset of the rows for decode(). We check that the nength at nulls() is no greater than that at decode().

Updates call sites of DecodedVector::nulls(). Updates tests.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8579

Reviewed By: bikramSingh91

Differential Revision: D53158552

Pulled By: oerling

fbshipit-source-id: 7431b5d72252b8cc3d2b69ff4e5c47f2092c5db3