[native] Prepare for the exchange protocol upgrade
This is the change needed on presto_cpp side to implement the new exchange
protocol.  Implement `ExchangeSource::getDataSizes` and use it when the header
`X-Presto-Get-Data-Size` exists (will be removed in the future once coordinator
support the `HEAD` calls), or the endpoint is called with HTTP `HEAD` method.

All changes here are backward-compatible.  The new protocol will not be used
until we start calling `getDataSizes` on velox side in `ExchangeClient`.

See https://github.com/prestodb/presto/issues/21926 for details about design of
the new protocol.

Also added new runtime stats `numTopOutputBuffers`.