Use splits weights. (#8822)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8822

HiveConnectorSplit has splitWeight member.
It is being used by Presto coordinator to schedule splits while monitoring how busy Tasks are with running and queued splits.
This change adds book keeping for such weights.
We update them only for TableScans, as this is what Presto Java does.
In addition, we add numRunningTableScanSplits and numQueuedTableScanSplits, which are also updated only for TableScans.

After this change is in, we will need to update presto_cpp to use and report the new stats.

Reviewed By: pranjalssh

Differential Revision: D54049212

fbshipit-source-id: 9dfeaa6885185e8aa94e4477dcee91c8932851a7