Use hash table for single map complex type subscript (#8798)
Summary:
Use hash table for single map complex type subscript

Deltoid NGA is limited by constantly sorting a map of 12K keys where
the key is an array. The trivial fix is to use a hash table instead of
sorting. Sorting 12K keys is ~ 12K * 14 comparisons for the sort and
then 1K * 14 comparisons for the binary search. Using a hash there are
a few compares for collisions in build and typically one in probing.

To do: Make the function stateful and avoid remaking the hash table.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8798

Reviewed By: kevinwilfong

Differential Revision: D53903153

Pulled By: oerling

fbshipit-source-id: 2e9c6b8341f01dbc82a570d334d10728fb3dbb36