Do not sort map entries in a RowContainer dependent column (#8806)
Summary:
If a Map column is not a key column but a dependent column of
a RowContainer, do not sort map entries during serialization.
Only key columns have to be sorted, since they are used for
comparision, ex. in the case of being the key of HashBuild.
It is assumed that a dependent column of unsorted Map could not
be called by RowContainerSerde::compare().

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8806

Reviewed By: oerling

Differential Revision: D53958725

Pulled By: gggrace14

fbshipit-source-id: 1b6bf0d5e78369e4bbc269349cc60a2ae49cbcbf