Fix OOM caused by aggregation calculation on DECIMAL type (#8859)
Summary:
For aggregate calculations of type `DECIMAL`,
`LongDecimalWithOverflowState` should be used to save the intermediate
results of each group. The current implementation incorrectly uses
`DecimalAggregate`.

`sizeof(DecimalAggregate)` is nearly 400 bytes, but
`sizeof(LongDecimalWithOverflowState)` is only 32 bytes, so the current
incorrect implementation will take up more than 10 times the memory.
Especially when there are many groups(such as TPCH-1TB Q18), it is easy
to cause out of memory (OOM).

It should be noted that although this is a bug, there is no correctness
problem for those queries that can be successfully executed, because
the size of `DecimalAggregate` is larger than
`LongDecimalWithOverflowState`, but more memory is used (memory that
should be used).

Fixes https://github.com/facebookincubator/velox/issues/8858

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8859

Reviewed By: Yuhta

Differential Revision: D54247168

Pulled By: mbasmanova

fbshipit-source-id: 639a899d7a18e00546c9d996f292081e4e9bd12d