Add support for individual file to check.py (#8844)
Summary:
Currently, "scripts/check.py" only works when it receives a directory,
it does nothing when it receives an individual file.  However, sometimes
we may want to only check an individual file (for example, we only
modified one file and want to check it quickly), and this patch adds
this capability to it.

With this patch, the following command can be used to fix the format
of an individual file:
```
  scripts/check.py format tree --fix path/to/a/file
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8844

Reviewed By: bikramSingh91

Differential Revision: D54208821

Pulled By: kgpai

fbshipit-source-id: bf4fd7ccf10b63a1dd73502d3d1de76462f40e04