Fix flakiness in 2 tests in TableWriterArbitrationTest
Summary:
This fixes flakiness in these tests that may be encountered if the test finishes
before the Tasks get a chance to be destroyed. If this happens then when the
memory manager gets deleted, it can lead to errors because some memory
pools (which are maintained by the potentially active Tasks) would still exist
holding onto memory whereas it expects all to be destoryed. The fix here is
to simply wait for the Tasks to be deleted.

Reviewed By: xiaoxmeng

Differential Revision: D54226837

fbshipit-source-id: 199bdb71776829623670c63e76002a3a9e60dd65