PrestoSerializer should check the checksum before anything else
Summary:
Today, when deserializing, PrestoSerializer reads the header, allocates a RowVector of size `numRows` and then checks the
checksum.

We should check the checksum before allocating the RowVector.  Admittedly, `numRows` isn't included in the checksum so
local corruptions of `numRows` can't be detected, but if the serialized data is complete garbage `numRows` could be huge
and the MemoryPool could end up OOM'ing.  By checking the checksum first we get a clearer error message that better
points to the true issue.

Reviewed By: xiaoxmeng

Differential Revision: D54284594

fbshipit-source-id: 5e6ef87f789eeb9688d2c85b0676908aef7ccf49