Add support for returning results as Velox Vectors to ReferenceQueryRunner (#8880)
Summary:
**What**
This PR adds support for the ReferenceQueryRunner to return results in Velox vector format. To do this we add function supportsVeloxVectorResults which returns true if the  Query Runner supports returning results in RowVector format. At the moment only PrestoQueryRunner supports returning the results of executing an SQL query as Velox vectors.

**Why**
Adding support for getting the results as a Velox Vector will allow us to call the compare() api for certain nondeterministic functions that have custom result verifiers. This will allow us to compare results where the output of the reference db differs from Velox for e.g in cases where the ordering is different, and ensure the results of the function are correct.
Currently we are only able to do this for sorted input for certain plans (Single aggregation); with the above support we will be able to do it for unsorted input also.

**Note**
The support to use this in the aggregation fuzzer will follow in another PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8880

Reviewed By: mbasmanova

Differential Revision: D54338280

Pulled By: kgpai

fbshipit-source-id: 11b940e8819867323cd3b440076467238a15608e