Allow entry delimiter to be last in the input of SplitToMap().
Summary:
We detected inconsitency of SplitToMap function compared to Presto.
Presto can handle input string with the entry delimiter being the last
character, assuming the previous chars were parsed all right.
Velox throws in such cases.
Examples: 'ds=2024/' and '=', where '/' is the entry delimiter and '=' is
the key-value delimiter.

So, in this change we bring Velox to behave like Presto.

Reviewed By: gggrace14, kewang1024

Differential Revision: D54315225

fbshipit-source-id: 09c796d600aa40b2ae8fb7a701650987dff3c9a6