Clear and fix a bug in Task::noMoreSplits().
Summary:
'noMoreSplits' message could come for nodes in both grouped and ungrouped execution mode.
We moved out all split promises to outside of the lock to fulfill.
In grouped execution case we might have multiple split stores with unfulfilled promises.
Moving them out like this in a loop can cause some promises to be destroyed unfulfilled:
```
splitPromises = std::move(it.second.splitPromises);
```
This can cause hanging Drivers and zombie Tasks.

In addition, the code is a bit confusing.

In this change we fix the bug and make code more clear.

Note: I believe such situation (with grouped execution) cannot happen, since coordinator
will likely send 'no more splits for group' message first and acknowledge it before sending
'no more splits' message, which means 'no more split groups'.
This way we would never really have any outstanding promises to fulfill in 'no more splits' message
for a grouped execution node.
But this is an assumption and better to have it covered in the code, in case it is not the case and
messages can get into a race.

Reviewed By: xiaoxmeng

Differential Revision: D54319594

fbshipit-source-id: 876bd52bb876cb11bd42384c2a5c8a133cc573c1