Move folly init and gtest init before registration of functions. (#8882)
Summary:
Move the folly init and gtest initialization before registration of any functions. We need to do this as certain gflags  are used to determine which set of aggregate functions to register, for e.g, when running the Aggregation fuzzer with presto as source of truth we only want to register presto supported aggregate signatures. The gflags are only set when folly::init is called and as flags are uninitialized we would only register signatures not supported in Presto.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8882

Reviewed By: xiaoxmeng, kagamiori

Differential Revision: D54338227

Pulled By: kgpai

fbshipit-source-id: a3c22a19073cc2db7bddb27d21c6f8776dd10391