Upgrade Boost to 1.84.0 (#8679)
Summary:
The boost version 1.84.0 fixes a bug that prevents
the addition of a group of inverse functions.
The fix was provided in Boost PR
https://github.com/boostorg/math/pull/1007

The update occurs for the setup scripts for CentOS and Ubuntu. MacOS already installs the
latest available Boost version on brew.

In addition, the requirement for the Boost version is updated to make sure the proper version is used.
There is no known build problem when upgrading the version.
Fixes issue: https://github.com/facebookincubator/velox/issues/8624

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8679

Reviewed By: Yuhta

Differential Revision: D54384316

Pulled By: kgpai

fbshipit-source-id: 7442b028a9b2f1b36b3190a7909a4a55ef52e72c