Add CAST(real as decimal) (#8575)
Summary:
Supports CAST(real as decimal) and makes the result more precise by taking
below two consideration.

1. The significant decimal digits precision of double and float includes the
integral part, therefore when scaling up, the integral size needs to be
excluded from the scale factor. This PR takes integral digits into
consideration in rescaleFloatingPoint to avoid introduing unexpected digits to
the result.
E.g. in `cast(99999.99 as decimal(38, 18))`, 99999.99 * 1e15 is
`99999990000000008192` which is not accurate. But if we take integral digits
into consideration, and mutliply 99999.99 with 1e10 before rounding,
`999999900000000` is got.

2. When the given scale is less than the precision provided by double or float,
the input value is scaled up to the max precision before rounding, and scaled
down to the required scale.
E.g. in `cast(0.95 as decimal(4, 1))`, `(long double)0.95 * 10` is
`9.499999999` which is rounded to 9. But if we firstly scale it up by 1e6,
`950000` can be got, and after dividing 1e5 and rounding, 10 is got as the final
result.

https://github.com/facebookincubator/velox/issues/8412

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8575

Reviewed By: amitkdutta

Differential Revision: D54408046

Pulled By: mbasmanova

fbshipit-source-id: c03e120b6f9f289fef799393cc595d8ef988bfbf